[ style ] 에러 아이콘 함수 분리하기
